+++ Makefile
@@ -1,12 +1,36 @@
-install:
-	@go get github.com/spf13/viper
-	@go get gorm.io/gorm
-	@go get gorm.io/driver/postgres
-	@go get github.com/gin-gonic/gin
+# 프로젝트 변수 설정
+PROJECT_NAME := todo-server
+PKG := cmd/main.go
+BINARY_NAME := $(PROJECT_NAME)
 
+# Go 관련 변수 설정
+GOBASE := $(shell pwd)
+GOBIN := $(GOBASE)/bin
+GOCMD := go
+
+# 빌드 작업 정의
 build:
-	@go build -o bin/app -v ./cmd/main.go
+	$(GOCMD) build -o $(BINARY_NAME) $(PKG)
 
+# 테스트 작업 정의
+test:
+	$(GOCMD) test -v ./...
 
-run:
-	@go run ./cmd/main.go
+# 실행 작업 정의
+run: build
+	./$(BINARY_NAME)
+
+# 의존성 설치 작업 정의
+install_deps:
+	$(GOCMD) mod download
+
+# 실행 파일 삭제 작업 정의
+clean:
+	rm -f $(BINARY_NAME)
+
+# 실행 파일 설치 작업 정의
+install: build
+	mv $(BINARY_NAME) $(GOBIN)
+
+# 모든 작업을 수행하는 기본 작업 정의
+all: install_deps build test
