+++ cmd/main.go
@@ -4,11 +4,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/gin-gonic/gin"
-	"github.com/potato/simple-restful-api/cmd/config"
 	"github.com/potato/simple-restful-api/infra/command"
+	"github.com/potato/simple-restful-api/infra/config"
 	"github.com/potato/simple-restful-api/infra/model"
-	"github.com/potato/simple-restful-api/infra/projector"
-	"github.com/potato/simple-restful-api/infra/query"
+	"github.com/potato/simple-restful-api/infra/monitoring"
+	"github.com/potato/simple-restful-api/infra/projector/watcher"
+	"github.com/potato/simple-restful-api/pkg"
 	"github.com/potato/simple-restful-api/pkg/rest/command"
 	"github.com/potato/simple-restful-api/pkg/rest/query"
 	"log"
@@ -22,18 +23,24 @@ var (
 func main() {
 	cfg := loadCfg()
 
+	// init event channel
 	eventChannel := make(chan eventModel.Event, 100)
 
-	eventStore := command.CreateCommander(cfg.Command, &eventChannel)
-	entityStore := query.CreateQuerier(cfg.Query)
+	// start event store
+	eventStore := command.CreateCommander(cfg.Command, eventChannel) // events -> event store
+	todoStore := todo.CreateTodoService(cfg.Query)                   // events -> entity -> pkg store
 
-	projector.InitProjector(cfg.Projector, &eventChannel)
-	go projector.RunConsumer(eventStore, entityStore, cfg.Projector.Kafka)
+	generator := todo.NewEntityGenerator(todoStore)                             // events -> entity
+	go watcher.CreateProjector(cfg.Projector, eventChannel).Run()               // event channel -> kafka
+	go watcher.StartEntityGenerator(eventStore, generator, cfg.Projector.Kafka) // kafka -> events -> projector
 
 	engine := gin.Default()
 	todoGroup := engine.Group("")
+
 	commandApi.NewTodoRouter(todoGroup, eventStore)
-	queryApi.NewTodoRouter(todoGroup, entityStore)
+	queryApi.NewTodoRouter(todoGroup, todoStore)
+
+	engine.Use(monitoring.LoggingMiddleware())
 
 	if err := engine.Run(fmt.Sprintf(":%s", cfg.Server.Port)); err != nil {
 		logger.Fatalln("Error running server", err)
