+++ infra/command/commander.go
@@ -1,7 +1,7 @@
 package command
 
 import (
-	"github.com/potato/simple-restful-api/cmd/config"
+	"github.com/potato/simple-restful-api/infra/config"
 	"github.com/potato/simple-restful-api/infra/db"
 	"github.com/potato/simple-restful-api/infra/model"
 	"gorm.io/gorm"
@@ -13,7 +13,7 @@ var (
 	logger = log.New(os.Stdout, "commander: ", log.LstdFlags|log.Lshortfile)
 )
 
-func CreateCommander(cfg *config.Command, ec *chan eventModel.Event) EventStore {
+func CreateCommander(cfg *config.Command, ec chan eventModel.Event) EventStore {
 	eventStoreDB, err := db.InitPostgresOrExit(cfg.EventStoreDB)
 	if err != nil {
 		logger.Fatalln("Error connecting to database", err)
@@ -22,7 +22,7 @@ func CreateCommander(cfg *config.Command, ec *chan eventModel.Event) EventStore 
 	return eventStore
 }
 
-func createEventStoreOrExit(eventStoreDB *gorm.DB, ec *chan eventModel.Event) EventStore {
+func createEventStoreOrExit(eventStoreDB *gorm.DB, ec chan eventModel.Event) EventStore {
 	// auto migration for EventStore
 	if err := eventStoreDB.Migrator().DropTable(&Event{}); err != nil {
 		logger.Fatalln("Error dropping table", err)
