+++ infra/projector/generator/mustbeimple.go
@@ -1,0 +1,7 @@
+package generator
+
+import "github.com/potato/simple-restful-api/infra/command"
+
+type EntityGenerator interface {
+	CreateEntityAnsSave(events []*command.Event) error
+}
