--- pkg/domain/spec/domain_command.go
@@ -1,13 +1,0 @@
-package spec
-
-type CreateTodo struct {
-	Title string `binding:"required"`
-}
-
-type UpdateTitle struct {
-	Title string `binding:"required"`
-}
-
-type UpdateStatus struct {
-	Status string `binding:"required"`
-}
