+++ pkg/domain/todospec/domain_command.go
@@ -1,0 +1,13 @@
+package todospec
+
+type CreateTodo struct {
+	Title string `binding:"required"`
+}
+
+type UpdateTitle struct {
+	Title string `binding:"required"`
+}
+
+type UpdateStatus struct {
+	Status string `binding:"required"`
+}
