+++ pkg/rest/query/router.go
@@ -2,11 +2,12 @@ package queryApi
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/potato/simple-restful-api/infra/query"
+	"github.com/google/uuid"
+	"github.com/potato/simple-restful-api/pkg/repository"
 	"strconv"
 )
 
-func NewTodoRouter(group *gin.RouterGroup, ets query.EntityStore) TodoRouter {
+func NewTodoRouter(group *gin.RouterGroup, ets repository.TodoStore) TodoRouter {
 	router := &todoRouter{
 		ets: ets,
 	}
@@ -23,14 +24,13 @@ type TodoRouter interface {
 }
 
 type todoRouter struct {
-	ets query.EntityStore
+	ets repository.TodoStore
 }
 
 func (t *todoRouter) GetAllByUserNo(c *gin.Context) {
 	// get param from path
-	userNo := c.Param("userNo")
-	userNoUint, err := strconv.ParseUint(userNo, 10, 64)
-	if err != nil {
+	userNo, err := t.parsePathUserNo(c)
+	if err != nil && userNo == 0 {
 		c.JSON(400, gin.H{
 			"error": err.Error(),
 		})
@@ -38,7 +38,7 @@ func (t *todoRouter) GetAllByUserNo(c *gin.Context) {
 	}
 
 	//TODO implement me
-	list, err := t.ets.GetTodosByUserNo(uint(userNoUint))
+	list, err := t.ets.GetTodosByUserNo(userNo)
 	if err != nil {
 		return
 	}
@@ -48,6 +48,47 @@ func (t *todoRouter) GetAllByUserNo(c *gin.Context) {
 }
 
 func (t *todoRouter) GetByUserNoAndID(c *gin.Context) {
-	//TODO implement me
-	panic("implement me")
+	userNo, id, err := t.parsePathParams(c)
+	if err != nil {
+		c.JSON(400, gin.H{
+			"error": err.Error(),
+		})
+		return
+	}
+	todo, err := t.ets.GetTodoByUserNoAndId(userNo, id)
+	if err != nil {
+		return
+	}
+
+	c.JSON(200, todo)
+}
+
+func (t *todoRouter) parsePathUserNo(c *gin.Context) (userNo uint, err error) {
+	userNo, err = t.parseUserNo(c)
+	if err != nil {
+		return 0, err
+	}
+
+	return userNo, nil
+}
+
+func (t *todoRouter) parsePathParams(c *gin.Context) (userNo uint, aggregateID uuid.UUID, err error) {
+	userNo, err = t.parseUserNo(c)
+	if err != nil {
+		return 0, uuid.Nil, err
+	}
+	aggregateID, err = uuid.Parse(c.Param("id"))
+	if err != nil {
+		return userNo, uuid.Nil, err
+	}
+
+	return userNo, aggregateID, nil
+}
+
+func (t *todoRouter) parseUserNo(c *gin.Context) (uint, error) {
+	userNoUint64, err := strconv.ParseUint(c.Param("userNo"), 10, 64)
+	if err != nil {
+		return 0, err
+	}
+	return uint(userNoUint64), nil
 }
